% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicciopasticcio-package.R
\docType{package}
\name{cicciopasticcio-package}
\alias{cicciopasticcio}
\alias{cicciopasticcio-package}
\title{cicciopasticcio: Pasticci a gogo}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Progetto test per uso laims.analysis template in cui fare un sacco di pasticci (a gogo, appunto).
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

}
\keyword{internal}
