% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(.data_path)
}
\arguments{
\item{.data_path}{(chr) path to the raw data to import}
}
\value{
a tibble
}
\description{
Function to import the data
}
