% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relevant_computation}
\alias{relevant_computation}
\title{Do what matters}
\usage{
relevant_computation(db)
}
\arguments{
\item{db}{(data.frame) data source for the relevant computation}
}
\value{
(num) our super amazing result
}
\description{
Do relevant computation only (but...test them!)
}
\examples{
relevant_computation(mtcars)
}
